+++ .github/workflows/kvm_watcher.yml
@@ -27,13 +27,12 @@ jobs:
           sudo apt install linux-tools-$(uname -r)
           sudo apt install linux-cloud-tools-$(uname -r)
           sudo modprobe kvm
-      - name: Run kvm_exits
+      - name: Run kvm_watcher
         run: |
-          cd eBPF_Supermarket/kvm_watcher/vm_exits
+          cd eBPF_Supermarket/kvm_watcher/
           make
-          sudo ./kvm_exits -t 5
-      - name: Run kvm_vcpu
-        run: |
-          cd eBPF_Supermarket/kvm_watcher/vcpu
-          make
-          sudo ./kvm_vcpu -w -t 5
+          sudo ./kvm_watcher -w -t 5
+          sudo ./kvm_watcher -e -t 5 -s
+          sudo ./kvm_watcher -n -t 5 
+          make clean
+
