--- eBPF_Supermarket/kvm_watcher/vcpu/Makefile
@@ -1,23 +1,0 @@
-ARCH ?= $(shell uname -m | sed 's/x86_64/x86/' \
-			 | sed 's/arm.*/arm/' \
-			 | sed 's/aarch64/arm64/' \
-			 | sed 's/ppc64le/powerpc/' \
-			 | sed 's/mips.*/mips/' \
-			 | sed 's/riscv64/riscv/' \
-			 | sed 's/loongarch64/loongarch/')
-APP = kvm_vcpu
-
-all: $(APP)
-
-.PHONY: $(APP)
-$(APP):
-ifeq ($(wildcard ./vmlinux.h),)
-	bpftool btf dump file /sys/kernel/btf/vmlinux format c > vmlinux.h
-endif
-	clang -g -O2 -target bpf -D__TARGET_ARCH_$(ARCH) -I/usr/include/x86_64-linux-gnu -I. -c $@.bpf.c -o $@.bpf.o
-	bpftool gen skeleton $@.bpf.o > $@.skel.h
-	clang -g -O2 -Wall -I . -c $@.c -o $@.o
-	clang -Wall -O2 -g $@.o -static -lbpf -lelf -lz -o $@
-
-clean:
-	rm -f *.o *.skel.h $(APP) vmlinux.h
